Move serializers to proper file in sponsors app
